antigravity/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
torch==2.1.2
torchvision==0.16.2
numpy==1.26.3
Pillow==10.2.0
requests==2.31.0
shapely==2.0.2
pyproj==3.6.1
opencv-python-headless==4.9.0.80
pydantic==2.5.3
python-dotenv==1.0.1
